meta_forge 编译通过

+++ src/meta_forge/src/main.cpp
@@ -1,11 +1,13 @@
 import meta_forge.forge;
+import core.type;
+import meta_forge.forge_phase;
 
 // int main( int argc, const char** argv )
 int main()
 {
     Forge forge;
 
-    int argc_o = 3;
+    usize argc_o = 3;
     const char** argv_o = new const char*[argc_o];
     argv_o[0] = "D:/nayuki_yq/build/windows/x64/debug/meta_forge.exe"; // Program name
     argv_o[1] = "--compilations=D:/nayuki_yq/.vscode";
